requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.9.1
aiosignal==1.3.1
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bip-utils==2.9.3
bitarray==2.9.2
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.1
cbor2==5.6.4
cdp-sdk==0.10.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
cleo==2.1.0
click==8.1.7
coincurve==20.0.0
crashtest==0.4.1
crcmod==1.7
cryptography==43.0.1
cytoolz==0.12.3
distlib==0.3.9
dulwich==0.21.7
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.1.0
frozendict==2.4.6
frozenlist==1.4.1
gunicorn==23.0.0
hexbytes==1.2.1
html5lib==1.1
idna==3.10
installer==0.7.0
iso8601==2.1.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
keyring==24.3.1
lxml==5.3.0
MarkupSafe==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
peewee==3.17.8
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==4.25.5
ptyprocess==0.7.0
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyNaCl==1.5.0
pyproject_hooks==1.2.0
pyseto==1.7.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
RapidFuzz==3.10.1
regex==2024.9.11
replit==4.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
shellingham==1.5.4
six==1.16.0
soupsieve==2.6
stripe==11.2.0
tomlkit==0.13.2
toolz==0.12.1
trove-classifiers==2024.10.21.16
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
web3==7.2.0
webencodings==0.5.1
websockets==13.0.1
Werkzeug==3.1.3
xattr==1.1.0
yarl==1.11.1
yfinance==0.2.49
